split types to remove circular dependencies
